fix cocoapods version on travis
otherwise defaults to 1.0.0 apparently, which does not install due to ruby version
conflicts. Not sure how to change this in travis (since the error log says 2.2.2 is
required but travis seems to only support 2.2.1). Not an ideal fix, but should do
the trick for the time being. (0.34.4 was the last cocoapods version used on master)